feat(map): generate it by script
