+++ backend/requirements.txt
@@ -5,6 +5,5 @@ psycopg2-binary
 folium
 python-dotenv
 shapely
-logging
 pytest
 pytest-mock
