+++ launch.sh
@@ -1,2 +1,2 @@
-#docker compose down --volumes
+docker compose down --volumes
 docker compose up --build --detach
